ref: refactor service package in bot to bot
