+++ bot/src/main/java/ru/tinkoff/edu/java/bot/bot/Bot.java
@@ -1,4 +1,4 @@
-package ru.tinkoff.edu.java.bot.service;
+package ru.tinkoff.edu.java.bot.bot;
 
 import com.pengrad.telegrambot.UpdatesListener;
 import com.pengrad.telegrambot.model.Update;
