+++ bot/src/main/java/ru/tinkoff/edu/java/bot/bot/commands/ListCommand.java
@@ -1,4 +1,4 @@
-package ru.tinkoff.edu.java.bot.service.commands;
+package ru.tinkoff.edu.java.bot.bot.commands;
 
 import com.pengrad.telegrambot.model.Update;
 import com.pengrad.telegrambot.request.SendMessage;
