+++ bot/src/main/java/ru/tinkoff/edu/java/bot/bot/commands/UntrackCommand.java
@@ -1,25 +1,26 @@
-package ru.tinkoff.edu.java.bot.service.commands;
+package ru.tinkoff.edu.java.bot.bot.commands;
 
 import com.pengrad.telegrambot.model.Update;
 import com.pengrad.telegrambot.request.SendMessage;
 import ru.tinkoff.edu.java.bot.storage.LinkStorage;
 
-public class TrackCommand implements Command{
+public class UntrackCommand implements Command{
 
     private final LinkStorage storage;
 
-    public TrackCommand(LinkStorage storage) {
+    public UntrackCommand(LinkStorage storage) {
         this.storage = storage;
     }
 
+
     @Override
     public String command() {
-        return "/track";
+        return "/untrack";
     }
 
     @Override
     public String description() {
-        return "начать отслеживание ссылки\nвведите команду в формате: " + command() + " url";
+        return "прекратить отслеживание ссылки\nвведите команду в формате: " + command() + " url";
     }
 
     @Override
@@ -29,7 +30,7 @@ public class TrackCommand implements Command{
         if (split.length == 1)
             return new SendMessage(chatId, description());
         String url = split[1];
-        storage.trackLink(chatId, url);
-        return new SendMessage(chatId, "Ссылка теперь отслеживается");
+        storage.untrackLink(chatId, url);
+        return new SendMessage(chatId, "Ссылка больше не отслеживается");
     }
 }
