+++ bot/src/main/java/ru/tinkoff/edu/java/bot/bot/processor/TelegramMessageProcessor.java
@@ -1,12 +1,12 @@
-package ru.tinkoff.edu.java.bot.service.processor;
+package ru.tinkoff.edu.java.bot.bot.processor;
 
 import com.pengrad.telegrambot.model.Update;
 import com.pengrad.telegrambot.request.SendMessage;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
-import ru.tinkoff.edu.java.bot.service.commands.Command;
-import ru.tinkoff.edu.java.bot.service.commands.HelpCommand;
-import ru.tinkoff.edu.java.bot.service.commands.UnknownCommand;
+import ru.tinkoff.edu.java.bot.bot.commands.Command;
+import ru.tinkoff.edu.java.bot.bot.commands.HelpCommand;
+import ru.tinkoff.edu.java.bot.bot.commands.UnknownCommand;
 
 import java.util.ArrayList;
 import java.util.List;
