+++ bot/src/main/java/ru/tinkoff/edu/java/bot/configuration/BotConfiguration.java
@@ -3,10 +3,10 @@ package ru.tinkoff.edu.java.bot.configuration;
 import lombok.AllArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import ru.tinkoff.edu.java.bot.service.Bot;
-import ru.tinkoff.edu.java.bot.service.TelegramRefresherBot;
-import ru.tinkoff.edu.java.bot.service.processor.TelegramMessageProcessor;
-import ru.tinkoff.edu.java.bot.service.processor.UserMessageProcessor;
+import ru.tinkoff.edu.java.bot.bot.Bot;
+import ru.tinkoff.edu.java.bot.bot.TelegramRefresherBot;
+import ru.tinkoff.edu.java.bot.bot.processor.TelegramMessageProcessor;
+import ru.tinkoff.edu.java.bot.bot.processor.UserMessageProcessor;
 import ru.tinkoff.edu.java.bot.storage.InMemoryLinkStorage;
 import ru.tinkoff.edu.java.bot.storage.LinkStorage;
