+++ bot/src/main/java/ru/tinkoff/edu/java/bot/configuration/CommandConfig.java
@@ -3,7 +3,7 @@ package ru.tinkoff.edu.java.bot.configuration;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import ru.tinkoff.edu.java.bot.service.commands.*;
+import ru.tinkoff.edu.java.bot.bot.commands.*;
 import ru.tinkoff.edu.java.bot.storage.LinkStorage;
 
 @Configuration
