+++ scrapper/src/main/java/ru/tinkoff/edu/java/scrapper/dto/response/GithubResponse.java
@@ -5,7 +5,10 @@ import jakarta.validation.constraints.NotNull;
 
 import java.time.OffsetDateTime;
 
-public record GithubResponse(@NotNull @JsonProperty(value = "full_name") String fullName, @NotNull @JsonProperty(value = "pushed_at") OffsetDateTime pushedAt,
-                             @NotNull @JsonProperty(value = "updated_at") OffsetDateTime updatedAt) {
+public record GithubResponse(@NotNull @JsonProperty(value = "full_name") String fullName,
+                             @NotNull @JsonProperty(value = "pushed_at") OffsetDateTime pushedAt,
+                             @NotNull @JsonProperty(value = "updated_at") OffsetDateTime updatedAt,
+                             @NotNull @JsonProperty(value = "open_issues_count") Long openIssuesCount,
+                             @NotNull @JsonProperty(value = "forks") Long forksCount) {
 
 }
