+++ scrapper/src/main/java/ru/tinkoff/edu/java/scrapper/dto/response/StackOverflowResponse.java
@@ -5,5 +5,8 @@ import jakarta.validation.constraints.NotNull;
 
 import java.time.OffsetDateTime;
 
-public record StackOverflowResponse(@NotNull String link, @NotNull @JsonProperty(value = "last_activity_date") OffsetDateTime lastActivityDate) {
+public record StackOverflowResponse(@NotNull String link,
+                                    @NotNull @JsonProperty(value = "last_activity_date") OffsetDateTime lastActivityDate,
+                                    @NotNull @JsonProperty(value = "answer_count") Integer answerCount,
+                                    @JsonProperty(value = "comment_count") Integer commentCount) {
 }
