--- scrapper/src/main/java/ru/tinkoff/edu/java/scrapper/service/updater/LinkUpdater.java
@@ -1,5 +1,0 @@
-package ru.tinkoff.edu.java.scrapper.service.updater;
-
-public interface LinkUpdater {
-    int update();
-}
