+++ scrapper/src/main/java/ru/tinkoff/edu/java/scrapper/updater/LinkUpdater.java
@@ -1,0 +1,5 @@
+package ru.tinkoff.edu.java.scrapper.updater;
+
+public interface LinkUpdater {
+    int update();
+}
